refactor:商品类目接口调整

+++ mall-pms/pms-api/src/main/java/com/youlai/mall/pms/entity/PmsSpu.java
@@ -21,8 +21,8 @@ public class PmsSpu extends BaseEntity {
     private String name;
     private Long categoryId;
     private Long brandId;
-    private BigDecimal originPrice;
-    private BigDecimal price;
+    private Integer originPrice;
+    private Integer price;
     private Integer sale;
     private String pic;
     private String album;
